Added ColourControllerEditor
ColourControllerEditor currently fills out empty object references while in the editor so that they don't have to be done on Start()

ColourController now runs on 8 Colour values, with all values implemented onto current prefabs.

the next step is to make it auto-recolour in the editor and implement more themes, with easier ways to switch between them. This may be easier with a CanvasHandler Editor Script.